Abbreviate long field descriptors at write time so that they don't get clobbered at runtime:     dirtybuf -> dtbuf     desiredvnodes -> desvn     numvnodes -> numvn     freevnodes -> frevn The vmstats column has only 5 characters available for descriptors, but up to 13 were used.  The extras get clobbered at runtime by interrupt values and/or descriptors on systems with more than 12 interrupt sources.     %slo-z -> %sloz This one is in the "extended" vmstats area and doesn't get clobbered now.
Removed stale documentation of desvn.

Changed a descriptor:
    tfree -> totfr
so that it is consistent with the abbreviations for other free counts
(daefr and prcfr) and thus almost decodeable.

Fixed missing documentation of tfree/totfr.  This and everything else
in the extended vmstats area is misdocumented as being in a certain
place in the vmstats column.