Extend the keyboard character size to 24 bits.
Because we use an int to store keyboard chacacters and their flags, we
can easily store the flags in the top byte instead of the second byte.
This means it's a lot easier to make Unicode work.

The only change that still needs to be made, is that keyent_t's map is
extended to u_int.

Obtained from:	//depot/projects/newcons/sys/sys/kbio.h