1. New feature: --packages-build. This option will switch to --packages-only mode to install dependencies that are only used for the build of the "parent" port that is specified on the command line, but will build the ports on the local system for anything that is a runtime dependency, and the port itself. [1]
At this point it can only be used with one port on the command
line, multiport and -a support are "on the list," as is
compatibility with the -t option.

This new option required some more complex error handling for
the --package* options which has been broken out into its own
function.

2. Modify fail() to be less finicky about the input comments

3. Twiddle some whitespace

[1] Special funding committment for this feature and package
support generally by:	Modirum MDpay