Fix 2 places where we generate a list of ports by using cut to process directory names in /var/db/pkg.
The previous code assumed that the value of PKG_DBDIR would always
be 3 directories deep, which is not necessarily a valid assumption.

In one spot this allows us to avoid an external shell call altogether,
in the other an internal function is a necessary evil to avoid other
even more complex alternatives.