Add a new option, --always-fetch If that option is NOT used, check the local file system for the package we're looking for. If it's not available, or the option IS used, fetch it.
Move the code for package fetching into a function

If we don't get a rational value from grep'ing the list of available
package versions, and/or the "newest" available port seems to be older
than the latest version from the ports tree, try fetching the exact
version of $new_port by name. This solves the issue of (for example)
autoconf-2.1.3 vs. autoconf-2.6.2. The latter appears never to have
a current package if the former exists.

Move a comment to the section it directly applies to