1. Instead of having a pca() to describe the post-config action collect the stuff that always runs after config is done into a function, and use the pca() logic to determine what to print. This also restores the whitespace to consistency between the modes (one port, multiport, -a).
2. 9-CURRENT has its own package repo now, hurray!

3. Instead of fetching the directory listing for each port category
(devel, ports-mgmt, etc.) every time we need to check a port, fetch
it once and save it to a temporary file. This lets us do several
cool things:
	a. Save a lot of time not having to re-fetch each iteration
	b. Run the sed code to fix %2c -> , up front
	c. Add a sed pattern to fix %2b -> +
	d. Run a variety of different patterns to try and find the
	   latest_pv

4. Not finding a package (or even a package repo) should only fail()
if we are using -PP, not if we're just using -P.

5. Move the error message for "no package in -PP mode" to a variable
for both reuse and code readability.

6. We only want to run the logic tree on whether $latest_pv is up to
date or not if that variable has a value, so add appropriate tests.