New Feature: ============ At the end of the config phase print the list of what is about to be done, and prompt the user on whether to proceed or not. This requires some infrastructure in update_port() to keep the list of what to do, and in post_config() to do the printing and prompting. This needs refinement for multiports.
Internal Cleanups:
==================
1. Instead of saving some of the new variables unconditionally in
safe_exit(), move the things that are only done in the config phase
to that section, and move build_deps_il to the !config section.
2. Do a more thorough job with the terminal escape sequences. It
now works in -a mode (although not thoroughly tested). This required
some annoying twiddling of spaces and such with PM_DEPTH, but so far
it seems "ok."
3. Rename init_parent_port to init_term_printf which is more accurate
and less confusing.
4. Since $new_port may be set as a side effect of the new feature
above, conditionalize it when we get there in the "main" process.