First cut at merging some features from NetBSD's install: o add ability to generate mtree meta files based on what's installed. o add ability to set destdir, so the metadata in the above file can filter that o add some rudamentary support for generating digests of the files
there's still some problems with this code (sizes that are generated
seem to be wrong) and there's some additional pieces that are needed
from NetBSD's install as well as hacks to the build system.  We're
installing about 17k files now.  10k are recorded in the metadata
file, while 7k aren't.  The vast majority of which are links of some
flavor or another.

Pieces to merge still: nopriv install and support for using install to
manage links rather than using ln.

Also, ee installs its nls files to the wrong place, which will need to
be fixed.  The metadata shows this.