Crude hack to make LLVM build on FreeBSD/arm.
LLVM needs a function to flush the icache of certain memory regions.
glibc has a function for it called __clear_cache(), but FreeBSD's libc
doesn't have it. I don't care about flushing the icache at all, because
it's only used when JITting, which we are not going to do with Clang.