Migrate the AR9160 specific initialisation code to use the linux-named arrays rather than the FreeBSD HAL named arrays.
The Linux arrays are being left as untouched as possible for now to make it
easier to merge in updates to the ath9k/initvals.h.

Add "bank6tpc" init/setup logic from linux ath9k. I'm not 100% sure that
I've ported the changes over correctly but the code has been tested
on a 9160 in 11b/11g AP mode.

This commit seems to resolve the majority of the "missed beacon" issues
I was seeing when using the AR9160 on the Ubiqiti Routerstation Pro platform.
I'm still able to trigger them by busying the CPU/bus. I haven't yet
tested this in anything resembling heavy/varied production use.