First (fairly complete) version of changing the behavior of the -G option.
In the original version of portmaster the concept of 2 distinct passes
through the dependency tree, the first to do all the 'make config's
and the second to build everything was intrinsic to the design. When
released publicly there were a non-trivial number of users who objected
to the idea of having 'make config' be mandatory, so I added the -G
option to disable the whole first pass.  However as time went on, more
and more features were added that depended on the first pass, so it was
necessary to rethink how -G works.

In this version, the 2-pass design is now universal, and the only thing
-G does is suppress the actual OPTIONS dialog. This has several benefits:
1. Now all features are available to -G users
2. The new method of suppressing the OPTIONS dialog is 100% effective
   (thus it's no longer necessary to use BATCH to truly suppress them
3. The --force-config option now overrides -G so users who really hate
   OPTIONS can have -G in their rc file, but still have it available
   by using --force-config on the command line

Lots of variables, functions, and text have been modified to substitute
out references to "config" in favor of "first pass"

I'm going to finish realigning the whitespace in a different pass