Increase MINSIZE; there isn't much point in doing 512-byte transactions, as they will always be either perfectly aligned or short.
Increase STEP to MAXSIZE * 4; having STEP equal to MAXSIZE means that
the last round actually tests sequential reading / writing, which is
not what we're after.

Instead of testing all offsets from 0 to size, test 0 and power-of-two
multiples of BSIZE, since we assume that the physical block size is a
power-of-two multiple of BSIZE.

This should both shorten the test's run time and improve the
significance of its output.