In messages that refer to the INDEX file, use the variable to make it clear what we're working with.
Be more careful to check that there is a line for the port in the INDEX
when working --index-only.

In dependency_check() be clear when we're using the INDEX instead of ports

In the package fetching routine use a better string match to see if we have
already downloaded the directory listing. The previous one worked but would
match too many files for net* and x11*, especially when installing a lot of
ports. So add a '.' to differentiate these.