Don't call update_build_l() (which also updates PM_NEEDS_UPDATE) in multiport() or update_port(), instead always call it from the main routine, which occurs after the user has made a decision about any +IGNOREME files that might be present. (Also simplifies a bit.)
Related to the above, in multiport() if the child returns and the
port in question is not in PM_NEEDS_UPDATE (usually/[only?] because
there is an +IGNOREME and they chose not to proceed) then add the
port to PM_MULTI_BUILT and decrement $num_ports so that it will be
ignored in the build phase.

In post_first_pass() return if PM_NEEDS_UPDATE is empty. This is a
micro-op in the common case, but allows us to give the right message
to the user in the -a case where the only port that needs updating
has an +IGNOREME file, and the user chose not to proceed.

For -a add some whitespace around the "all ports are up to date"