Introduce get_answer() to handle processing of y/n questions that we want to be able to override on the command line with either -y or -n
Utilize this new function to provide -y|n support for the following:
-e, -s, --clean-distfiles, --clean-packages, --check-port-dbdir, and
--check-depends. This also allows me to deprecate the horribly hacky
--clean-distfiles-all and --clean-packages-all in favor of -y.

Sort usage() to group features that use -y|n support, which also
better reflects "do stuff" features vs. "list stuff"

Update recommendations to use --clean-distfiles to include [-y], not [-all]

For -s, if the user chooses not to delete the stale port flip the question
about continuing to list it as a dependency, and correspondingly flip the
default. This has no effect on -y (the condition is never reached) but
allows -n to work reliably, and preserving the same default action (to
delete the empty +REQUIRED_BY) reduces POLA.