Add an error check for -y and -n both supplied
Add get_answer_g(), a generic version of the previous get_answer()
Use this function to wrap all of the remaining cases of user input

Convert get_answer() to get_answer_yn() and call get_answer_g() if
we need to prompt the user, otherwise use the same logic to handle
the -y and -n options.
Update the callers of the former get_answer() to match.