Add necessary address space mapping. The Altix architecture has distinct address spaces for PCI devices and CPUs. That is, the S-Hub translates addresses on the FSB and coming from CPUs to addresses valid on the PCI bus. This is abstracted and implemented in FreeBSD by having the handle contain the CPU address, while the resource itself describes the PCI bus address. Since devices are mapped when resources are activated and the bus handle is derived from that, have the SGI PCI bus driver handle the activation of resources instead of passing it up to the nexus.
Note that the Altix architecture allows for IRQ values larger than 255 even
though the intline register in the PCI configuration space does not. FreeBSD
also doesn't support it by virtue of having an IRQ resource space that's
limited. On my machines the IRQ values are within bounds so I'm not going to
worry about this detail. It may very well be theoretical.

With this change the bge(4) driver attaches properly. The one thing stopping
us from netbooting is... interrupts...