Use a seperate driver for the SAL-based console on Altix. While initially uart(4) seemed like a good idea (and for the low-level console it wasn't a bad idea), the fact that the SAL-based console doesn't have any I/O port or memory mapped I/O resources and uart(4) expects and needs that means that uart(4) is simply not a good idea.
The new driver in its current form is almost functional complete. It only
lacks proper interrupt handling. The reason for that is that interrupts
aren't exactly working on Altix, so it's a bit hard to test. Also, SAL
only allows us to have receive interrupts, not transmit interrupts. As
such, a bit more effort is required to make sure the TX path doesn't
stall (read: polling and/or timeouts are needed).

The uart(4) driver changes are reverted.