Some fixes - fix BAW checking/update (which isn't entirely fixed, I'll be fiddling with this shortly!) and simplify the swq TX routine.
This is unfortunately one of those aggregate commits which I dislike
doing.

* Remove an unneeded mbuf argument to ath_tx_swq(). ath_buf should already
  be setup by now and bf->bf_m is the first mbuf in the chain.

* Now that NULL data QOS frames are not receiving sequence numbers from
  the TID, we can't check them against the BAW window when we're scheduling
  packets. Otherwise it's likely it'll be outside the BAW and will remain
  so, so the TX will hang until the watchdog timer kicks in.

The code duplication involved in checking whether a frame is going to be
given a sequence number and treated as part of the BAW is .. well, horrible.

Unfortunately, I didn't think about needing to also not call the BAW update
upon packet completion. That's going to have to be the next check.