Modify how/when the frame descriptor is being setup.
For packets being immediately queued to the hardware, setup the
descriptor before handing it off to the hardware.

For packets being queued to the per-TID software queue, defer
setting up the descriptor details (setup, rate control, chaining)
until the packet is about to be queued to the hardware.

This way, packet aggregation (RIFS burst, A-MPDU) can setup
the descriptor list as needed.

The only downside at the moment is that the rate control decision
is done (very) early on. I'll fix this at a later stage so software
retries can use different rates as needed.