I've been seeing situations where the TX queue has validly completed packets in it. This shouldn't occur - the hardware should've sent an interrupt to signal that a TX completion had occured.
So on a hunch, I decided to shuffle the interrupt mitigation bits
up a bit, because I figured that the hardware may be sending a
TX mitigation interrupt but not asserting TXOK at the time.

This seems to be doing the "right" thing - ie, at the time the
mitigation timer expiry has occured, the AR_ISR_S0/AR_ISR_S1/AR_ISR_S2
registers seem to have the correct TX queue interrupt cause bits set;
as they would stay set (once set) until AR_ISR_RAC is read.

I'm going to have to chase this up with the MAC guys at Atheros to see
exactly what the correct behaviour is.

(Now there are still device timeout messages, but with an empty TX
 queue- this signifies a different problem.)