In find_and_delete_distfiles() (run after a successful install) change the default pattern to be foo-[0-9]* instead of foo* [1]
This matches 19,305 out of 25,106 unique distfiles, and dramatically
reduces false positives for files like foo-1.23 vs. foo-bar-2.34. Of the
ones that this pattern misses, 4,713 don't match the traditional distfile
format anyway, so it's a huge net win; and should make running without -D
much less annoying.

When the new pattern doesn't match any existing distfiles, fall back to
the old one. Do this by calling with the pattern, not the file name. This
simplifies processing in the function a bit.

Also in find_and_delete_distfiles(), cache matches to $DI_FILES so that
we don't have to re-test that same file again.

Change the method used to print the prompt in get_answer_g() to echo -e
so that I can be more flexible in formatting prompt messages.

Suggested by:	Lapo Luchini <lapo@lapo.it> [1]