Enable printf's where a BAR frame _would_ be TXed.
I'm doing this so I (and testers) can keep track of when BAR's
should be TXed. This way if traffic pauses are seen that coincide
with this BAR message, we can (hopefully!) assume that a BAR frame
exchange would've flushed things correctly.

This also will let me implement/debug the "correct" sending of BAR
frames - ie, sending them when the TID has been emptied rather than
on each failed software retry attempt. That could get slightly spammy.

There's a few places where BAR frames should likely be TX'ed,
specifically when a TID flush is done (and the node isn't being
deleted.) But that'll come in the next phase of work.