Put the txqactive and kickpcu flags behind ath_lock for now, to avoid racey access to them from different threads.
Since the ath_intr() code can be called at the same time as a taskqueue, we need to
ensure that these don't clash.

It doesn't happen on my testing on single-core MIPS boards w/out PREEMPTION enabled.
It's likely to happen on PREEMPTION kernels and with multi-core CPUs.

The real solution is to do what the reference driver does - create a pcu lock and
serialise all PCU access behind that. I'll look at doing this at a later time.