Improve the handling of the -F option with packages, especially with --packages-only:
In update_port() do better at saying what we're launching a child to do

In post_first_pass() make clear that we are fetch'ing, not building,
installing, etc.

In multiport() call check_fetch_only when we're done

In the main body add fetch'ed ports to the CUR_DEPS list so that they
are only fetched once.