Begin adding support to force a full chip reset when the reset HAL method is called, rather than a warm reset.
Merlin (at least) seems to get very annoyed after a while when interference
is present. It's possible that part of why the AR9220/AR9280 is getting angry
is due to (more) incorrect software programming of things, but once things
_get_ too angry, it stays that way through a warm reset.

Some further testing and resaerch is needed to figure out what's going on.
It's likely that the "correct" thing to do with this WAR is to extend the
reset HAL method to take a "warm or full" flag and then set a "full"
reset during a stuck beacon situation or when interference is detected.